fix(lint): preserve error cause in api.ts and split PageTitleContext exports
- api.ts: pass the caught error as `cause` when building the HTTP
  fallback error so the original (e.g. JSON parse) error is not lost
  (preserve-caught-error).
- PageTitleContext: move the non-component export out of the component
  file so react-refresh/only-export-components is satisfied; update all
  importers.
